python_algo/crm-marketing/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.0.1
python-dotenv==0.19.1
requests==2.26.0
sendgrid==6.9.0
twilio==7.0.0
